cups-print-system/requirements.txt
Flask
pycups
Pillow
